+++ Tutorial-o-Matic/zips/templates/Tutorial-o-Matic/tutorial_styles_template.mm
@@ -1,6 +1,8 @@
 <map version="freeplane 1.9.13">
 <!--To view this file, download free mind mapping software Freeplane from https://www.freeplane.org -->
-<node TEXT="New Mindmap" FOLDED="false" ID="ID_258304564" STYLE="oval"><hook NAME="MapStyle" background="#ffffff">
+<node TEXT="New Mindmap" FOLDED="false" ID="ID_258304564" STYLE="oval">
+<font SIZE="18"/>
+<hook NAME="MapStyle" background="#ffffff">
     <properties show_icon_for_attributes="true" edgeColorConfiguration="#808080ff,#ff0000ff,#0000ffff,#00ff00ff,#ff00ffff,#00ffffff,#7c0000ff,#00007cff,#007c00ff,#7c007cff,#007c7cff,#7c7c00ff" mapUsesOwnSaveOptions="true" save_modification_times="false" save_last_visited_node="default" show_note_icons="true" save_folding="default" fit_to_viewport="false"/>
 
 <map_styles>
@@ -41,60 +43,72 @@ code {
 <icon BUILTIN="edit"/>
 <richcontent TYPE="NOTE" CONTENT-TYPE="plain/markdown"/>
 </stylenode>
-<stylenode TEXT="ToM-Tutorial" ID="ID_1317696283" ICON_SIZE="20 pt" BACKGROUND_COLOR="#f3d0c5" STYLE="rectangle" BORDER_WIDTH="3 px">
+<stylenode TEXT="ToM-Tutorial" ID="ID_1317696283" ICON_SIZE="24 pt" COLOR="#000000" BACKGROUND_COLOR="#fbcbbc" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
+<cloud COLOR="#f2ded8" SHAPE="ROUND_RECT"/>
+<attribute NAME="ToM_TabLabel" VALUE=" "/>
 </stylenode>
-<stylenode TEXT="ToM_TOC" ID="ID_1160910212" ICON_SIZE="12 pt" BACKGROUND_COLOR="#f5ff95" STYLE="rectangle" BORDER_WIDTH="3 px">
-<icon BUILTIN="emoji-1F989"/>
-<icon BUILTIN="emoji-25B6"/>
-</stylenode>
-<stylenode TEXT="ToM_newPage" ID="ID_1643366420" ICON_SIZE="12 pt" BACKGROUND_COLOR="#95d9ff" STYLE="rectangle" BORDER_WIDTH="3 px">
+<stylenode TEXT="ToM_TOC" ID="ID_1160910212" ICON_SIZE="16 pt" COLOR="#000000" BACKGROUND_COLOR="#f5ff95" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="emoji-25B6"/>
 </stylenode>
-<stylenode TEXT="ToM_nextPage" ID="ID_318145687" ICON_SIZE="12 pt" BACKGROUND_COLOR="#ff9595" STYLE="rectangle" BORDER_WIDTH="3 px">
+<stylenode TEXT="ToM_newPage" ID="ID_1643366420" ICON_SIZE="16 pt" COLOR="#000000" BACKGROUND_COLOR="#95d9ff" STYLE="rectangle" BORDER_WIDTH="3 px">
+<icon BUILTIN="emoji-1F989"/>
+<icon BUILTIN="emoji-25B6"/>
+</stylenode>
+<stylenode TEXT="ToM_nextPage" ID="ID_318145687" ICON_SIZE="16 pt" COLOR="#000000" BACKGROUND_COLOR="#ff9595" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="emoji-25B6"/>
 </stylenode>
 <stylenode TEXT="ToM_note" ID="ID_1031812973" ICON_SIZE="16 pt" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="emoji-1F4F0"/>
+<cloud COLOR="#f9f9b9" SHAPE="ROUND_RECT"/>
 </stylenode>
 <stylenode TEXT="ToM_showMenu" ID="ID_723735694" ICON_SIZE="16 pt" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="emoji-1F5B1"/>
+<cloud COLOR="#f9f9b9" SHAPE="ROUND_RECT"/>
 </stylenode>
 <stylenode TEXT="ToM_goto" ID="ID_1953218925" ICON_SIZE="16 pt" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="emoji-2197"/>
+<cloud COLOR="#f9f9b9" SHAPE="ROUND_RECT"/>
 </stylenode>
 <stylenode TEXT="ToM_menuAction" ID="ID_1981660950" ICON_SIZE="16 pt" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="emoji-1F525"/>
+<cloud COLOR="#f9f9b9" SHAPE="ROUND_RECT"/>
 </stylenode>
 <stylenode TEXT="ToM_groovy" ID="ID_1578586059" ICON_SIZE="16 pt" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="emoji-1F951"/>
+<cloud COLOR="#f9f9b9" SHAPE="ROUND_RECT"/>
 </stylenode>
 <stylenode TEXT="ToM_copy" ID="ID_691299161" ICON_SIZE="16 pt" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="emoji-1F4CB"/>
+<cloud COLOR="#f9f9b9" SHAPE="ROUND_RECT"/>
 </stylenode>
 <stylenode TEXT="ToM_select" ID="ID_1792407381" ICON_SIZE="16 pt" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="emoji-2B55"/>
+<cloud COLOR="#f9f9b9" SHAPE="ROUND_RECT"/>
 </stylenode>
 <stylenode TEXT="ToM_openMap" ID="ID_323134366" ICON_SIZE="16 pt" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="mindmap"/>
+<cloud COLOR="#f9f9b9" SHAPE="ROUND_RECT"/>
 </stylenode>
 <stylenode TEXT="ToM_openTutMap" ID="ID_255978903" ICON_SIZE="16 pt" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="links/file/freeplane_mindmap"/>
+<cloud COLOR="#f9f9b9" SHAPE="ROUND_RECT"/>
 </stylenode>
 <stylenode TEXT="ToM_showNode" ID="ID_1304839137" ICON_SIZE="16 pt" STYLE="rectangle" BORDER_WIDTH="3 px">
 <icon BUILTIN="emoji-1F989"/>
 <icon BUILTIN="emoji-1F517"/>
+<cloud COLOR="#f9f9b9" SHAPE="ROUND_RECT"/>
 </stylenode>
 </stylenode>
 <stylenode LOCALIZED_TEXT="styles.AutomaticLayout" POSITION="right" STYLE="bubble">
@@ -124,6 +138,5 @@ code {
 </stylenode>
 </map_styles>
 </hook>
-<font SIZE="18"/>
 </node>
 </map>